block: sed-opal: provide cmd_start function
Every step is an OPAL_CALL and thus starts with a common sequence of
required steps like resetting the cmd buffer, construct the OPAL_CALL
and provide the uid and method bytestring. Instead of having those
actions spread across the file in all the step-functions, provide a
function that combines all those steps.

Signed-off-by: Jonas Rabenstein <jonas.rabenstein@studium.uni-erlangen.de>