block: sed-opal: reduce memory footprint
A sed-opal enabled device currently has a struct opal_dev allocated
independent of the fact whether sed-opal is ever used on this device. As
that struct has embedded buffers for sending and receiving data of
currently 2048 bytes each, the struct is a little bit more than 4k of
size. With a buddy allocator this will get placed in an 8k memory chunk.
Moreover, it is most likely that there is only a short period of
sed-opal command usage (e.g. unlocking at boot and locking on shutdown),
but the memory is claimed for the whole time.
By moving those buffers as well as further session state out of the
struct opal_dev the resident memory set can be reduced for the price of
a slightly increased delay on the execution of sed-opal commands due to
the memory allocation.